+++ src/main/java/br/com/zalf/prolog/webservice/gente/controleintervalo/model/TipoIntervalo.java
@@ -1,7 +1,7 @@
 package br.com.zalf.prolog.webservice.gente.controleintervalo.model;
 
-import br.com.zalf.prolog.webservice.colaborador.Cargo;
-import br.com.zalf.prolog.webservice.colaborador.Unidade;
+import br.com.zalf.prolog.webservice.colaborador.model.Cargo;
+import br.com.zalf.prolog.webservice.colaborador.model.Unidade;
 import com.google.gson.annotations.SerializedName;
 
 import java.sql.Time;