+++ src/main/java/br/com/zalf/prolog/webservice/gente/quiz/quizModelo/QuizModeloResource.java
@@ -1,6 +1,6 @@
 package br.com.zalf.prolog.webservice.gente.quiz.quizModelo;
 
-import br.com.zalf.prolog.webservice.colaborador.Cargo;
+import br.com.zalf.prolog.webservice.colaborador.model.Cargo;
 import br.com.zalf.prolog.webservice.commons.network.AbstractResponse;
 import br.com.zalf.prolog.webservice.commons.network.Response;
 import br.com.zalf.prolog.webservice.permissao.pilares.Pilares;
@@ -35,6 +35,13 @@ public class QuizModeloResource {
         return service.getModelosQuizByCodUnidade(codUnidade);
     }
 
+    @GET
+    @Secured
+    @Path("/{codUnidade}/{codModeloQuiz}/completos")
+    public ModeloQuiz getModeloQuiz(@PathParam("codUnidade") Long codUnidade,@PathParam("codModeloQuiz") Long codModeloQuiz) {
+        return service.getModeloQuiz(codUnidade, codModeloQuiz);
+    }
+
     @POST
     @Secured
     @Path("/{codUnidade}")
@@ -64,5 +71,4 @@ public class QuizModeloResource {
             return Response.error("Erro ao alterar as funções vinculadas ao quiz");
         }
     }
-
 }